+++ mayan/apps/checkouts/serializers.py
@@ -1,11 +1,11 @@
 from django.utils.translation import ugettext_lazy as _
 
-from rest_framework import serializers
 from rest_framework.generics import get_object_or_404
 
 from mayan.apps.acls.models import AccessControlList
 from mayan.apps.documents.models import Document
 from mayan.apps.documents.serializers.document_serializers import DocumentSerializer
+from mayan.apps.rest_api import serializers
 
 from .models import DocumentCheckout
 from .permissions import permission_document_check_out