+++ mayan/apps/messaging/serializers.py
@@ -1,8 +1,8 @@
 from django.utils.translation import ugettext_lazy as _
 
-from rest_framework import serializers
 from rest_framework.reverse import reverse
 
+from mayan.apps.rest_api import serializers
 from mayan.apps.rest_api.serializer_mixins import CreateOnlyFieldSerializerMixin
 from mayan.apps.rest_api.relations import FilteredPrimaryKeyRelatedField
 from mayan.apps.user_management.querysets import get_user_queryset