+++ mayan/apps/rest_api/fields.py
@@ -1,7 +1,7 @@
 from django.utils.module_loading import import_string
 from django.utils.translation import ugettext_lazy as _
 
-from rest_framework import serializers
+from . import serializers
 
 
 class DynamicSerializerField(serializers.ReadOnlyField):