+++ mayan/apps/rest_api/relations.py
@@ -4,10 +4,11 @@ from django.db.models import Manager
 from django.db.models.query import QuerySet
 from django.utils.translation import ugettext_lazy as _
 
-from rest_framework import serializers
 from rest_framework.relations import HyperlinkedIdentityField
 
 from mayan.apps.common.utils import resolve_attribute
+
+from . import serializers
 
 
 class FilteredRelatedFieldMixin: